+++ History.md
@@ -9,6 +9,7 @@
 * added more Lytics options, by @slindberg and @araddon
 * added trait alias to trak.io
 * added MouseStats, by @Koushan
+* added Tapstream, by @adambard
 
 0.11.11 / 2013-8-9
 ------------------